+++ src/server/partition/mod.rs
@@ -1,10 +1,9 @@
-use std::{error::Error, time::Duration};
-
+use super::{Request, Response};
 use async_trait::async_trait;
 
 #[async_trait(?Send)]
 pub trait Partition {
-    type Error: Error;
+    type Error: std::error::Error;
 
     async fn serve_idempotent(&self, request: Request) -> Result<Response, Self::Error>;
 
@@ -22,32 +21,6 @@ pub trait PartitionCreator<P: Partition> {
 pub struct PartitionId {
     pub topic: String,
     pub partition_number: u64,
-}
-
-pub enum Request {
-    RemoveExpired { expiry_duration: Duration },
-    CreatePartition,
-    RemovePartition,
-
-    Read { offset: u64 },
-    LowestOffset,
-    HighestOffset,
-
-    Append { record_bytes: Vec<u8> },
-}
-
-pub type Record = crate::commit_log::Record<'static>;
-
-pub enum Response {
-    Read { record: Record, next_offset: u64 },
-    LowestOffset(u64),
-    HighestOffset(u64),
-
-    Append { write_offset: u64 },
-
-    ExpiredRemoved,
-    PartitionCreated,
-    PartitionRemoved,
 }
 
 pub mod in_memory;
